+++ js/utils/get-random-positive-float.js
@@ -1,10 +1,8 @@
-function getRandomWholeNum(randomMinNum, randomMaxNum) {
-  if (randomMinNum < randomMaxNum && randomMinNum >= 0) {
-    randomMinNum = Math.ceil(randomMinNum);
-    randomMaxNum = Math.floor(randomMaxNum);
-    return Math.floor(Math.random() * (randomMaxNum - randomMinNum + 1)) + randomMinNum;
+function getRandomFractionNum(randomMinNum, randomMaxNum, digits) {
+  if (randomMinNum < randomMaxNum && randomMinNum >= 0 && digits >= 0) {
+    return (Math.random() * (randomMaxNum - randomMinNum) + randomMinNum).toFixed(digits);
   }
   return 'Error';
 }
 
-export default getRandomWholeNum;
+export default getRandomFractionNum;
