+++ js/utils/get-random-positive-integer.js
@@ -1,8 +1,10 @@
-function getRandomFractionNum(randomMinNum, randomMaxNum, digits) {
-  if (randomMinNum < randomMaxNum && randomMinNum >= 0 && digits >= 0) {
-    return (Math.random() * (randomMaxNum - randomMinNum) + randomMinNum).toFixed(digits);
+function getRandomWholeNum(randomMinNum, randomMaxNum) {
+  if (randomMinNum < randomMaxNum && randomMinNum >= 0) {
+    randomMinNum = Math.ceil(randomMinNum);
+    randomMaxNum = Math.floor(randomMaxNum);
+    return Math.floor(Math.random() * (randomMaxNum - randomMinNum + 1)) + randomMinNum;
   }
   return 'Error';
 }
 
-export default getRandomFractionNum;
+export default getRandomWholeNum;
